Avoid silent overflow for numpy compatibility

+++ harp/io.py
@@ -90,7 +90,7 @@ def read(
         micros = np.ndarray(nrows, dtype=np.uint16, buffer=data, offset=payloadoffset, strides=stride)
         payloadoffset += 2
         time = micros * _SECONDS_PER_TICK + seconds
-        payloadtype = payloadtype & ~0x10
+        payloadtype = payloadtype & ~np.uint8(0x10)
         if epoch is not None:
             time = epoch + pd.to_timedelta(time, "s")  # type: ignore
         index = pd.Series(time)
